requirements.txt
pylint
discord.py
